feat: added AuthSuccessPage. Refactor Routes
