+++ frontend/heartmage-candles/src/App.tsx
@@ -1,18 +1,12 @@
-import { Routes, Route } from 'react-router-dom';
-
 import { AuthProvider } from './context/AuthContext';
 import Header from './components/shared/Header';
-import PrivateRoutes from './routes/PrivateRoutes';
-import PublicRoutes from './routes/PublicRoutes';
+import Routes from './routes/Routes';
 
 function App() {
   return (
     <AuthProvider>
       <Header />
-      <Routes>
-        <Route path="*" element={<PublicRoutes />} />
-        <Route path="admin/*" element={<PrivateRoutes />} />
-      </Routes>
+      <Routes />
     </AuthProvider>
   );
 }
