+++ frontend/heartmage-candles/src/context/AuthContext.tsx
@@ -1,7 +1,4 @@
-import { FC, createContext, ReactNode, useState, useEffect } from 'react';
-import { useNavigate } from 'react-router-dom';
-import { AuthHelper } from '../helpers/AuthHelper';
-import { AuthApi } from '../services/AuthApi';
+import { FC, createContext, ReactNode, useState } from 'react';
 
 export interface AuthProviderProps {
   children?: ReactNode;
@@ -21,59 +18,6 @@ const AuthContext = createContext<IAuthContext>(initialValue);
 
 const AuthProvider: FC<AuthProviderProps> = ({ children }) => {
   const [isAuth, setIsAuth] = useState(initialValue.isAuth);
-  const navigate = useNavigate();
-
-  useEffect(() => {
-    const token = AuthHelper.getToken();
-    if (token) {
-      setIsAuth(true);
-    } else {
-      navigate('/auth');
-    }
-  }, []);
-
-  useEffect(() => {
-    const intervalInMinutes = 2.83;
-    const intervalInMilliseconds = intervalInMinutes * 60 * 1000;
-
-    const isTokenExpiringSoon = (expireAt: Date) => {
-      const timeLeft = new Date(expireAt).getTime() - Date.now();
-      return timeLeft < 5 * 60 * 1000;
-    };
-
-    const refreshAuthToken = async () => {
-      const token = AuthHelper.getToken();
-
-      if (token && isTokenExpiringSoon(token?.ExpireAt)) {
-        try {
-          const tokenRequest = {
-            accessToken: token.AccessToken,
-            refreshToken: token.RefreshToken,
-          };
-
-          const newToken = await AuthApi.refreshToken(tokenRequest);
-
-          if (newToken.data) {
-            AuthHelper.setToken(newToken.data);
-            console.log('Token refreshed successfully.');
-          } else {
-            throw new Error('Failed to refresh token.');
-          }
-        } catch (error) {
-          console.error('Error refreshing token:', error);
-          navigate('/auth');
-        }
-      }
-    };
-
-    refreshAuthToken();
-
-    const interval = setInterval(() => {
-      refreshAuthToken();
-    }, intervalInMilliseconds);
-
-    return () => clearInterval(interval);
-  }, []);
 
   return (
     <AuthContext.Provider value={{ isAuth, setIsAuth }}>
