+++ frontend/heartmage-candles/src/routes/PrivateRoutes.tsx
@@ -1,45 +1,68 @@
-import { FC } from 'react';
-import { Routes, Route } from 'react-router-dom';
+import { FC, useContext, useEffect } from 'react';
+import { Outlet, useNavigate } from 'react-router-dom';
 
-import Navbar from '../components/admin/Navbar';
-import AllCandlePage from '../pages/admin/AllCandlePage';
-import CandleDetailsPage from '../pages/admin/CandleDetailsPage';
-import AllDecorsPage from '../pages/admin/AllDecorPage';
-import AllLayerColorPage from '../pages/admin/AllLayerColorPage';
-import LayerColorPage from '../pages/admin/LayerColorPage';
-import AllSmellPage from '../pages/admin/AllSmellPage';
-import SmellPage from '../pages/admin/SmellPage';
-import AllWickPage from '../pages/admin/AllWickPage';
-import WickPage from '../pages/admin/WickPage';
-import BotPage from '../pages/admin/BotPage';
-import NotFoundPage from '../pages/home/NotFoundPage';
-
-import Style from './PrivateRoutes.module.css';
+import { AuthContext } from '../context/AuthContext';
+import { AuthHelper } from '../helpers/AuthHelper';
+import { AuthApi } from '../services/AuthApi';
 
 const PrivateRoutes: FC = () => {
-  return (
-    <>
-      <div className={Style.AdminContent}>
-        <Navbar />
-        <main className={Style.AdminContainer}>
-          <Routes>
-            <Route index element={<AllCandlePage />} path="candles" />
-            <Route element={<CandleDetailsPage />} path="candles/:id" />
-            <Route index element={<AllDecorsPage />} path="decors" />
-            <Route element={<AllDecorsPage />} path="decors/:id" />
-            <Route index element={<AllLayerColorPage />} path="layerColors" />
-            <Route element={<LayerColorPage />} path="layerColors/:id" />
-            <Route index element={<AllSmellPage />} path="smells" />
-            <Route element={<SmellPage />} path="smells/:id" />
-            <Route index element={<AllWickPage />} path="wicks" />
-            <Route element={<WickPage />} path="wicks/:id" />
-            <Route element={<BotPage />} path="bot" />
-            <Route path="*" element={<NotFoundPage />} />
-          </Routes>
-        </main>
-      </div>
-    </>
-  );
+  const { setIsAuth } = useContext(AuthContext);
+
+  const navigate = useNavigate();
+
+  useEffect(() => {
+    const intervalInMinutes = 2.83;
+    const intervalInMilliseconds = intervalInMinutes * 60 * 1000;
+
+    const isTokenExpiringSoon = (expireAt: Date) => {
+      const timeLeft = new Date(expireAt).getTime() - Date.now();
+      return timeLeft < 5 * 60 * 1000;
+    };
+
+    const refreshAuthToken = async () => {
+      const token = AuthHelper.getToken();
+
+      if (token && isTokenExpiringSoon(token?.ExpireAt)) {
+        try {
+          const tokenRequest = {
+            accessToken: token.AccessToken,
+            refreshToken: token.RefreshToken,
+          };
+
+          const newToken = await AuthApi.refreshToken(tokenRequest);
+
+          if (newToken.data) {
+            AuthHelper.setToken(newToken.data);
+            console.log('Token refreshed successfully.');
+          } else {
+            throw new Error('Failed to refresh token.');
+          }
+        } catch (error) {
+          console.error('Error refreshing token:', error);
+          navigate('/auth');
+        }
+      }
+    };
+
+    refreshAuthToken();
+
+    const interval = setInterval(() => {
+      refreshAuthToken();
+    }, intervalInMilliseconds);
+
+    return () => clearInterval(interval);
+  }, []);
+
+  useEffect(() => {
+    const token = AuthHelper.getToken();
+    if (token) {
+      setIsAuth(true);
+    } else {
+      navigate('/auth');
+    }
+  }, []);
+
+  return <Outlet />;
 };
 
 export default PrivateRoutes;
